+++ CHANGES.md
@@ -2,7 +2,9 @@
 
 #### 0.2.4
 * Updates 'rapidminer-release' plugin to version 0.1.6
-* Updates 'rapidminer-publish' plugin to version 0.2.0
+* Replaces 'rapidminer-publish' plugin by 'maven-publish' plugin
+* Adds JUnit, nebula-test and Spock as test dependencies
+* Changes release tasks to build and publish
 
 #### 0.2.3
 * Updates 'rapidminer-code-quality' plugin to version 0.2.1